Combined BrimstoneChestOpenable and BrimstoneChest
